fix: 스키마 가격 필드 타입 정합성 강화
- RouteResponse.min_price: float→Decimal 변경 + ge=0 제약 추가
- route_service: 불필요한 float() 변환 제거
- PriceHistoryResponse: min/max/avg_price에 ge=0 제약 추가

+++ backend/app/services/route_service.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timedelta, timezone
+from decimal import Decimal
 
 from sqlalchemy import select, or_, case, literal, func
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -73,9 +74,9 @@ class RouteService:
             )
             .group_by(FlightPrice.route_id)
         )
-        price_map: dict[int, float] = {}
+        price_map: dict[int, Decimal] = {}
         for pr in price_result.all():
-            price_map[pr.route_id] = float(pr.min_price)
+            price_map[pr.route_id] = pr.min_price
 
         responses = []
         for row in rows:
